requirements.txt
langchain
langchain-openai
streamlit
requests
python-dotenv 